// packages/dashboard/src/http/types/ListTasks.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export const listTasksStatus200DataStatusEnum = {
    TODO: "TODO",
    IN_PROGRESS: "IN_PROGRESS",
    DONE: "DONE"
} as const;

export type ListTasksStatus200DataStatusEnumKey = (typeof listTasksStatus200DataStatusEnum)[keyof typeof listTasksStatus200DataStatusEnum];

export type ListTasksStatus200 = {
    data: {
        id: string;
        title: string;
        description: string | null;
        status: ListTasksStatus200DataStatusEnumKey;
        position: number;
        /**
         * @description
         * Format: `date`
         * @type string
        */
        dueDate: string | null;
        /**
         * @description
         * Format: `date-time`
         * @type string
        */
        createdAt: string;
    }[];
};

export type ListTasksOptions = {
    body?: never;
    path?: never;
    query?: never;
    headers?: never;
};

export type ListTasksResponses = {
    "200": ListTasksStatus200;
};

/**
 * @description Union of all possible responses
*/
export type ListTasksResponse = ListTasksStatus200;